# openapi.yaml
openapi: 3.1.0
info:
  title: Serpent.io API
  description: >
    Backend contract for the Snake Social Club frontend. Mirrors the
    `Services` interface in `frontend/src/services/types.ts` (auth, games,
    leaderboard). This file is the source of truth for the backend
    implementation.
  version: 0.1.0
  license:
    name: UNLICENSED

servers:
  - url: /api/v1

tags:
  - name: auth
    description: Sign up, sign in, and session management.
  - name: games
    description: Playing and spectating live game runs.
  - name: leaderboard
    description: Per-mode high scores.

security: []

paths:
  /auth/signup:
    post:
      operationId: signUp
      tags: [auth]
      summary: Create a new player and start a session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "201":
          description: Account created and signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: >
            Invalid username/password, or the username is already taken.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/signin:
    post:
      operationId: signIn
      tags: [auth]
      summary: Sign in with an existing username and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          description: Invalid username or password.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/signout:
    post:
      operationId: signOut
      tags: [auth]
      summary: End the current session.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Signed out.
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/session:
    get:
      operationId: getSession
      tags: [auth]
      summary: Resolve the session for the given token.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The session for the presented token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          description: Missing or invalid token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /games:
    post:
      operationId: createGame
      tags: [games]
      summary: Start a new game run for the signed-in player.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [mode, gridSize]
              properties:
                mode:
                  $ref: "#/components/schemas/GameMode"
                gridSize:
                  type: integer
                  minimum: 1
      responses:
        "201":
          description: The newly created game.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSnapshot"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /games/active:
    get:
      operationId: listActiveGames
      tags: [games]
      summary: List every game currently in progress, for spectating.
      responses:
        "200":
          description: Active games, highest score first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameSnapshot"

  /games/{gameId}:
    parameters:
      - name: gameId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getGame
      tags: [games]
      summary: Fetch a single game snapshot, for spectating.
      responses:
        "200":
          description: The game snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSnapshot"
        "404":
          description: No game with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateGame
      tags: [games]
      summary: Publish the latest state of the caller's own game run.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [snake, food, score, status]
              properties:
                snake:
                  type: array
                  items:
                    $ref: "#/components/schemas/Point"
                food:
                  $ref: "#/components/schemas/Point"
                score:
                  type: integer
                  minimum: 0
                status:
                  $ref: "#/components/schemas/GameStatus"
      responses:
        "200":
          description: The updated game snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSnapshot"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The game belongs to a different player.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No game with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /leaderboard/scores:
    post:
      operationId: submitScore
      tags: [leaderboard]
      summary: Record a finished run's score for the signed-in player.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [mode, score]
              properties:
                mode:
                  $ref: "#/components/schemas/GameMode"
                score:
                  type: integer
                  minimum: 0
      responses:
        "201":
          description: The recorded score entry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      operationId: topScores
      tags: [leaderboard]
      summary: Top scores for one game mode, one entry per player.
      parameters:
        - name: mode
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/GameMode"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        "200":
          description: >
            Each player's best score for this mode, highest first, oldest
            first on ties. A player is never listed more than once.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoreEntry"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: The `token` from a `Session`, as returned by signup/signin.

  responses:
    Unauthorized:
      description: No session, or the token is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        message:
          type: string

    GameMode:
      type: string
      enum: [walls, pass-through]

    GameStatus:
      type: string
      enum: [active, over]

    Point:
      type: object
      additionalProperties: false
      required: [x, y]
      properties:
        x:
          type: integer
        y:
          type: integer

    User:
      type: object
      additionalProperties: false
      required: [id, username]
      properties:
        id:
          type: string
        username:
          type: string

    Session:
      type: object
      additionalProperties: false
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    Credentials:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6

    ScoreEntry:
      type: object
      additionalProperties: false
      required: [id, userId, username, mode, score, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        mode:
          $ref: "#/components/schemas/GameMode"
        score:
          type: integer
        createdAt:
          type: integer
          description: Unix epoch milliseconds.

    GameSnapshot:
      description: >
        The authoritative snapshot of one player's run, as seen by
        spectators.
      type: object
      additionalProperties: false
      required:
        - id
        - userId
        - username
        - mode
        - gridSize
        - snake
        - food
        - score
        - status
        - startedAt
        - updatedAt
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        mode:
          $ref: "#/components/schemas/GameMode"
        gridSize:
          type: integer
        snake:
          type: array
          items:
            $ref: "#/components/schemas/Point"
        food:
          $ref: "#/components/schemas/Point"
        score:
          type: integer
        status:
          $ref: "#/components/schemas/GameStatus"
        startedAt:
          type: integer
          description: Unix epoch milliseconds.
        updatedAt:
          type: integer
          description: Unix epoch milliseconds.